feat: refactor Share List into a separate component
